Move CSV to separate module
