+++ extensions/csv/src/main/resources/META-INF/services/com.hazelcast.jet.pipeline.file.impl.ReadFileFnProvider
@@ -1,0 +1,1 @@
+com.hazelcast.jet.csv.CsvReadFileFnProvider
