+++ extensions/pom.xml
@@ -37,6 +37,7 @@
         <module>cdc-debezium</module>
         <module>cdc-mysql</module>
         <module>cdc-postgres</module>
+        <module>csv</module>
         <module>elasticsearch/elasticsearch-5</module>
         <module>elasticsearch/elasticsearch-6</module>
         <module>elasticsearch/elasticsearch-7</module>
