+++ hazelcast-jet-core/pom.xml
@@ -122,12 +122,6 @@
             <artifactId>jackson-core</artifactId>
             <version>${jackson.jr.version}</version>
         </dependency>
-        <!-- TODO we probably need to move the file formats to own module -->
-        <dependency>
-            <groupId>com.fasterxml.jackson.dataformat</groupId>
-            <artifactId>jackson-dataformat-csv</artifactId>
-            <version>${jackson.jr.version}</version>
-        </dependency>
         <dependency>
             <groupId>javax.cache</groupId>
             <artifactId>cache-api</artifactId>
